Make doc_structure_checker runnable
