+++ check.sh
@@ -7,3 +7,7 @@ cargo check --workspace --all-targets
 #cargo clippy --workspace --all-targets -- -D warnings -W clippy::all
 cargo clippy -- -D warnings -W clippy::all
 cargo test --features "logging, toml_config" --workspace --doc
+
+# Check doc structure defination
+chmod +x scripts/doc_structure_checker
+scripts/doc_structure_checker
